+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/floating-toolbar/SaveFragmentCompositionModal.js
@@ -56,7 +56,7 @@ const SaveFragmentCompositionModal = ({
 
 	return (
 		<ClayModal
-			className="page-editor__fragment-composition__modal"
+			className="page-editor__save-fragment-composition__modal"
 			observer={observer}
 			size="lg"
 		>
