+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/ContainerWithControls.js
@@ -26,8 +26,8 @@ import duplicateItem from '../../thunks/duplicateItem';
 import {useSelectItem} from '../Controls';
 import Topper from '../Topper';
 import FloatingToolbar from '../floating-toolbar/FloatingToolbar';
+import SaveFragmentCompositionModal from '../floating-toolbar/SaveFragmentCompositionModal';
 import Container from './Container';
-import SaveFragmentCompositionModal from './SaveFragmentCompositionModal';
 
 const ContainerWithControls = React.forwardRef(
 	({children, item, layoutData}, ref) => {
